pavelganaratsky BOT 5 TG/requirements.txt
aiogram
python-dotenv
aiosqlite
aiohttp-socks
